+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - You can now open a playlist folder and see all it's tracks
 - Support for .opus files with multiple pictures
 - Support showing and removing specific pictures for tracks with multiple pictures
+- Remember opened sidebar folders when the app relaunches
 - Update button design
 - Trap focus in modals
 - Store electron data in new `space.kasper.ferrum` app data folder